profile 테이블 추가 생성 및 회원탈퇴 로직 변경

+++ src/app/auth/withdrawal/route.ts
@@ -1,5 +1,6 @@
 import { Database } from "@/types/supabase";
 import { createClient } from "@supabase/supabase-js";
+import { NextResponse } from "next/server";
 
 export async function POST(request: Request) {
   const { user_id } = await request.json();
@@ -14,9 +15,22 @@ export async function POST(request: Request) {
       },
     }
   );
+  //프로필 삭제부터
+  const { error: profileDeleteError } = await supabase
+    .from("profile")
+    .delete()
+    .eq("id", user_id);
+
+  if (profileDeleteError) {
+    console.log(profileDeleteError);
+    return NextResponse.json(profileDeleteError, { status: 400 });
+  }
+
   const { data, error } = await supabase.auth.admin.deleteUser(user_id);
+
   if (error) {
-    return new Response("fail", { status: 400 });
+    console.log(error);
+    return NextResponse.json(error, { status: 400 });
   }
 
   if (!error) {
@@ -31,7 +45,7 @@ export async function POST(request: Request) {
       await Promise.all([deleteLikes(), deleteComments()]);
       return new Response("success", { status: 200 });
     } catch (error) {
-      return new Response("fail", { status: 400 });
+      NextResponse.json(error, { status: 400 });
     }
   }
 }
